pfake calculation

+++ src/main/java/blockchain/core/Transaction.java
@@ -33,9 +33,9 @@ public class Transaction implements Serializable
     public ArrayList<TransactionInput> inputs = new ArrayList<TransactionInput>();
     public ArrayList<TransactionOutput> outputs = new ArrayList<TransactionOutput>();
 
-    public int numberOfVerification = 0;
+    public double numberOfVerification = 0;
 
-    public int numberOfConfirmation = 0;
+    public double numberOfConfirmation = 0;
 
     public Transaction(byte[] privateKey, byte[] sender, byte[] recipient, float value, Schedule schedule, ArrayList<TransactionInput> inputs)
     {
